refactored tools, improve modularity
